Create separate files for camera
