+++ src/Playground/CMakeLists.txt
@@ -3,8 +3,10 @@ project(Playground)
 
 add_custom_target(copy_playground_data ALL COMMAND ${CMAKE_COMMAND} -E copy_directory ${CMAKE_SOURCE_DIR}/data ${CMAKE_CURRENT_BINARY_DIR}/data)   
 
+#Could also consider 'glob'-ing it here to prevent a FetchContent call
 set(sourceFiles
     Main.cpp
+    Camera.cpp
     PlaygroundApplication.cpp
 )
