+++ src/Playground/include/PlaygroundApplication.hpp
@@ -13,14 +13,16 @@
 #include <array>
 #include <optional>
 
-//Fwog Stuff
+#include <Camera.hpp>
 
+//Fwog Stuff
 #include <Fwog/BasicTypes.h>
 #include <Fwog/Buffer.h>
 #include <Fwog/Pipeline.h>
 #include <Fwog/Rendering.h>
 #include <Fwog/Shader.h>
 #include <Fwog/Texture.h>
+
 
 
 namespace Primitives
@@ -142,30 +144,6 @@ struct Skybox
     static Fwog::Texture MakeTexture();
 };
 
-struct Camera
-{
-    Camera();
-    void Update();
-
-    glm::vec3 camPos = glm::vec3(3.0f, 3.0f, 3.0f);
-    glm::vec3 target = glm::vec3(0.0f, 0.0f, 0.0f);
-    glm::vec3 up = glm::vec3(0.0f, 1.0f, 0.0f);
-
-    float nearPlane = 0.01f;
-    float farPlane = 5000.0f;
-    
-    struct CameraUniforms {
-        glm::mat4 viewProj;
-        glm::vec3 eyePos;
-    };
-
-    CameraUniforms cameraStruct;
-
-    std::optional<Fwog::TypedBuffer<CameraUniforms>> cameraUniformsBuffer;
-
-    //Skybox doesn't pass in the translation
-    std::optional<Fwog::TypedBuffer<CameraUniforms>> cameraUniformsSkyboxBuffer;
-};
 
 //Could potentially rename it
 struct GameObject
